Reduced code duplication, simplified imports
