+++ src/spec/assertion/toBeIn.spec.ts
@@ -1,7 +1,8 @@
-import { test, testSuite, expect, ftestSuite } from '../../testyCore';
+import { test, testSuite, expect } from '../../testyCore';
 
 @testSuite('Expect ToBeIn Tests')
 export class ExpectToMatch {
+
     @test(`'a' to be in, should succeed`)
     private aToBeInSuccess() {
         expect.toBeIn('a', ['a', 'b', 'c']);
