+++ src/spec/decorators/testSuiteTestsBase.ts
@@ -1,8 +1,8 @@
-import { beforeEach } from '../../lib/decorators/beforeEach.decorator';
 import { Report } from '../../lib/reporting/report/report';
 import { TestSuite } from '../../lib/tests/testSuite';
 import { TestVisitor } from '../../lib/tests/visitors/testVisitor';
 import { TestRunnerVisitor } from '../../lib/tests/visitors/testRunnerVisitor';
+import { beforeEach } from '../../lib/decorators/afterAndBefore.decorator';
 
 export class TestSuiteTestsBase {
     protected visitor: TestVisitor<Report>;
