+++ src/spec/tests/testRunnerVisitor.spec.ts
@@ -1,20 +1,19 @@
-import { testSuite, ftestSuite } from '../../lib/decorators/testSuite.decorator';
-import { test } from '../../lib/decorators/test.decorator';
-import { beforeEach } from '../../lib/decorators/beforeEach.decorator';
+import { testSuite, beforeEach, test, expect } from '../../testyCore';
+import { TestRunnerVisitor } from '../../lib/tests/visitors/testRunnerVisitor';
 import { TestSuite } from '../../lib/tests/testSuite';
 import { Test } from '../../lib/tests/test';
 import { TestStatus } from '../../lib/testStatus';
-import { expect } from '../../lib/assertion/expect';
 import { CompositeReport } from '../../lib/reporting/report/compositeReport';
 import { SuccessfulTestReport } from '../../lib/reporting/report/successfulTestReport';
-import { Report } from '../../lib/reporting/report/report';
-import { LeafReport } from '../../lib/reporting/report/leafReport';
 import { FailedTestReport } from '../../lib/reporting/report/failedTestReport';
 import { SkippedTestReport } from '../../lib/reporting/report/skippedTestReport';
-import { TestRunnerVisitor } from '../../lib/tests/visitors/testRunnerVisitor';
+import { Report } from '../../lib/reporting/report/report';
+import { LeafReport } from '../../lib/reporting/report/leafReport';
+
 
 @testSuite('Test Runner Visitor Tests')
 export class TestRunnerVisitorTests {
+
     private testRunnerVisitor: TestRunnerVisitor;
 
     @beforeEach()
