+++ src/testyCore.ts
@@ -2,8 +2,5 @@ export { test, ftest, xtest } from './lib/decorators/test.decorator';
 export { testSuite, ftestSuite, xtestSuite } from './lib/decorators/testSuite.decorator';
 export { expect } from './lib/assertion/expect';
 export { TestCase } from './lib/testCase';
-export { afterAll } from './lib/decorators/afterAll.decorator';
-export { afterEach } from './lib/decorators/afterEach.decorator';
-export { beforeEach } from './lib/decorators/beforeEach.decorator';
-export { beforeAll } from './lib/decorators/beforeAll.decorator';
+export { afterAll, afterEach, beforeEach, beforeAll } from './lib/decorators/afterAndBefore.decorator';
 export { TestResult } from './lib/reporting/report/testResult';
